Refactor choices list in ApplicationDetails to its own component
